TOOLS: move TreeBuilder
