+++ base/src/main/java/org/aya/concrete/stmt/QualifiedID.java
@@ -4,14 +4,15 @@ package org.aya.concrete.stmt;
 
 import kala.collection.Seq;
 import kala.collection.immutable.ImmutableSeq;
-import org.aya.util.error.SourcePos;
 import org.aya.generic.Constants;
+import org.aya.util.binop.SourceNode;
+import org.aya.util.error.SourcePos;
 import org.jetbrains.annotations.NotNull;
 
 public record QualifiedID(
   @Override @NotNull SourcePos sourcePos,
   @NotNull ImmutableSeq<@NotNull String> ids
-) {
+) implements SourceNode {
   public QualifiedID(@NotNull SourcePos sourcePos, @NotNull String id) {
     this(sourcePos, ImmutableSeq.of(id));
   }
